Add pipe mode, fix x64 offset handling, fix Hide column visibility
Server:
- Add 'pipe' launch mode: skips console allocation and listener socket,
  emits ---MYSEQ-READY--- / ---MYSEQ-END--- sentinels, suppresses prompts.
  Allows offset_wizard.py to drive the debug server via stdin/stdout pipes.
- Add 'sph' debug command: silent spawn walk (no per-spawn output, 10k limit)
  for quickly locating SpawnHeaderAddr without flooding output.
- Fix scanForPtr: was checking pPrev != 0 (never scanned); corrected to == 0,
  step by 8 bytes instead of 1, use dynamic base address instead of hardcoded.
- Fix NetworkServer::current_offset return type UINT -> QWORD so x64 addresses
  above 4GB are not truncated when compared in EQGameScanner.
- MemReader: fall back to 0x140000000 with a warning if base address detection
  returns 0 instead of silently using a null base.
- EQGameScanner: refactored ScanExecutable/ScanSecondary, modern C++ cleanup.
- myseqserver.ini: updated with current patch offsets and full debug CLI docs.

Client:
- Fix Hide/invisibility column not updating for non-pet spawns. The column
  update was gated on OwnerID != 0; moved outside that condition with a
  change-detection check so all spawn types reflect the current invis state.

+++ server/EQGameScanner.h
@@ -34,9 +34,9 @@ public:
 
 	void setExe(TCHAR* str);
 
-	DWORD findEQPointerOffset(DWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask);
+	QWORD findEQPointerOffset(DWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask);
 
-	DWORD findEQStructureOffset(DWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask, const QWORD baseEQPointerAddress);
+	QWORD findEQStructureOffset(DWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask, const QWORD baseEQPointerAddress);
 
 	bool ScanExecutable(HWND hDlg, IniReaderInterface* ir_intf, NetworkServerInterface* net_intf, bool write_out = false);
 
@@ -45,8 +45,8 @@ public:
 private:
 	bool compareData(PBYTE data, PBYTE byteMask, PCHAR charMask);
 	void updateFileInfoSection(std::ostringstream& outputStream, IniReaderInterface* ir_intf, bool write_out);
-	DWORD findAndProcessOffset(HWND hDlg, const std::string& section, const std::string& entry, IniReaderInterface* ir_intf, NetworkServerInterface* net_intf, std::ostringstream& outputStream, bool write_out);
-	void handleMatchResult(HWND hDlg, DWORD matchAddr, NetworkServerInterface* net_intf, int offsetType, const std::string& offsetName, IniReaderInterface* ir_intf, std::ostringstream& outputStream, bool write_out, bool& reload);
+	QWORD findAndProcessOffset(HWND hDlg, const std::string& section, const std::string& entry, IniReaderInterface* ir_intf, NetworkServerInterface* net_intf, std::ostringstream& outputStream, bool write_out);
+	void handleMatchResult(HWND hDlg, QWORD matchAddr, NetworkServerInterface* net_intf, int offsetType, const std::string& offsetName, IniReaderInterface* ir_intf, std::ostringstream& outputStream, bool write_out, bool& reload);
 
 	std::string executablePath;
 };
