+++ server/NetworkServer.h
@@ -30,7 +30,7 @@
 class NetworkServerInterface
 {
 public:
-	virtual UINT current_offset(int type) = 0;
+	virtual QWORD current_offset(int type) = 0;
 };
 
 class NetworkServer : public NetworkServerInterface
@@ -130,7 +130,7 @@ public:
 
 	string getCharName(MemReaderInterface* mr_intf);
 
-	UINT current_offset(int type);
+	QWORD current_offset(int type);
 
 private:
